No implicit pchar to string_view conversions (might lead to redundant std::strlen and buffer traversals).

+++ src/tests/gmsobjtests.cpp
@@ -179,7 +179,7 @@ void runStressTest( const std::string_view caption, f getelem )
    static_assert( card < std::numeric_limits<int>::max() );
    std::array<int, card> nums {};
    std::random_device rd;
-   int sum{};
+   uint64_t sum{};
    for( int k{}; k < ntries; k++ )
    {
       T shlst;
